Rework tests to appease Android Instrumented Tests
